Artist login remake

+++ src/main/java/map/project/musiclibrary/service/ArtistUserService.java
@@ -31,13 +31,21 @@ public class ArtistUserService {
         this.albumService = albumService;
     }
 
-    public ArtistUser addArtist(String name, String birthdateStr) throws ParseException {
-        SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
-        Date birthdate = dateFormat.parse(birthdateStr);
+    public ArtistUser addArtist(UserSession userSession, String name, String birthdateStr) {
+        if (userSession.isLoggedIn() && userSession.getCurrentUser().isAdmin()) {
+            SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+            Date birthdate;
+            try {
+                birthdate = dateFormat.parse(birthdateStr);
+            } catch (ParseException e) {
+                throw new IllegalArgumentException("Invalid birthdate format. Please use yyyy-MM-dd.");
+            }
 
-        ArtistUser artist = (ArtistUser) CreatorUserFactory.createCreatorUser("artist", name, birthdate);
+            ArtistUser artist = (ArtistUser) CreatorUserFactory.createCreatorUser("artist", name, birthdate);
 
-        return artistUserRepository.save(artist);
+            return artistUserRepository.save(artist);
+        }
+        throw new SecurityException("Only admin can add an artist");
     }
 
     public ArtistUser save(ArtistUser artistUser) {
@@ -48,49 +56,51 @@ public class ArtistUserService {
         return artistUserRepository.findByName(name).stream().findFirst().orElse(null);
     }
 
-    public List<ArtistUser> findAll() {
-        return artistUserRepository.findAll();
+    public List<ArtistUser> findAll(UserSession userSession) {
+        if (userSession.isLoggedIn() && userSession.getCurrentUser().isAdmin()) {
+            return artistUserRepository.findAll();
+        }
+        throw new SecurityException("Only admin can list all artists");
     }
 
     public Optional<ArtistUser> findById(Long id) {
         return artistUserRepository.findById(id);
     }
 
-    public List<NormalUser> getFollowers(String artistIdStr) throws NumberFormatException {
-        Long artistId = Long.parseLong(artistIdStr);
-        Optional<ArtistUser> artistUserOptional = findById(artistId);
-        if (artistUserOptional.isPresent()) {
-            ArtistUser artist = artistUserOptional.get();
-            return artist.getFollowers();
-        } else {
+    public List<NormalUser> getFollowers(UserSession userSession, String artistIdStr) throws NumberFormatException {
+        if (userSession.isLoggedIn() && userSession.getCurrentUser().isAdmin()) {
+            Long artistId = Long.parseLong(artistIdStr);
+            Optional<ArtistUser> artistUserOptional = findById(artistId);
+            if (artistUserOptional.isPresent()) {
+                ArtistUser artist = artistUserOptional.get();
+                return artist.getFollowers();
+            }
             throw new EntityNotFoundException("User with specified id not found");
         }
+        throw new SecurityException("Only admin can list all followers");
     }
 
     @Transactional
     public void delete(UserSession userSession, String idStr) throws NumberFormatException {
-        Long id = Long.parseLong(idStr);
-        Optional<ArtistUser> optional = artistUserRepository.findById(id);
-        if (optional.isPresent()) {
-            ArtistUser artist = optional.get();
-
-            // remove the link between the artist and label
-            artist.getLabel().getArtists().remove(artist);
-
-            // remove songs associated with the artist
-            // TODO - dependenta circulara fml
-            for (Song song : artist.getSongs()) {
-                songService.delete(song.getId().toString());
+        if (userSession.isLoggedIn() && userSession.getCurrentUser().isAdmin()) {
+            Long id = Long.parseLong(idStr);
+            Optional<ArtistUser> optional = artistUserRepository.findById(id);
+            if (optional.isPresent()) {
+                ArtistUser artist = optional.get();
+                // remove the link between the artist and label
+                artist.getLabel().getArtists().remove(artist);
+                // remove songs associated with the artist
+                for (Song song : artist.getSongs()) {
+                    songService.delete(song.getId().toString());
+                }
+                // remove albums associated with the artist
+                for (Album album : artist.getAlbums()) {
+                    albumService.delete(userSession, album.getId().toString());
+                }
+                artistUserRepository.deleteById(id);
             }
-
-            // remove albums associated with the artist
-            for (Album album : artist.getAlbums()) {
-                albumService.delete(userSession, album.getId().toString());
-            }
-
-            artistUserRepository.deleteById(id);
-        } else {
             throw new EntityNotFoundException("Artist was not found");
         }
+        throw new SecurityException("Only admin can delete an artist");
     }
 }
